# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Gen AI capture the flag with Go",
        "contact": {
            "name": "Artur Shad Nik"
        },
        "version": "1.0"
    },
    "host": "gobot-backend-ew7jmfmn3a-uw.a.run.app",
    "paths": {
        "/chat/{id}": {
            "get": {
                "description": "handler to fetch all past messages between the user and the bot",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Difficulty Level",
                        "name": "level",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "Get a response to a question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Chat with a bot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message",
                        "name": "message",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Difficulty Level",
                        "name": "level",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ConvoMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ConvoMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}